feat(Collection): add extra functionality to ObservableList
The ObservableList component now has additional functionality like
being able to check if the collection contains a given item and
set an element value at a specific index.

The fields have now been converted to properties and a new CurrentIndex
field has been added that allows methods that require an index to
still function if called via the limitations of a UnityEvent (which
only allows one parameter type).