fix(Utility): prevent error when attempting to pick no component
The TypePickerWindow would throw an error if a type was searched for
that didn't yield any results and then the window was still clicked
to perform a null selection. This would result in a selected index of
-1 and therefore create an out of bounds error.

This has been fixed by checking if the selectionIndex is not -1. It's
not worth checking if the index is within the upper bounds as the
list wouldn't allow that case to arise.