fix(Type): ensure SerializableType updates when assembly name changes
The internal assemblyQualifiedTypeName field within SerializableField
can be changed via reflection (and in the TypePickerAttributeDrawer)
but when the field data is changed, it never actually returned a new
serialized type when the ActualType property was called because
the setting of the internal private field wasn't ever checked for
updates.

This fix now stores the current value of the qualified name and will
create a new SerializedType if that value has changed. If the value
hasn't changed then it uses the current cached value to prevent
needing to do the `Type.GetType(string)` lookup.

A test has been added for the SerializableType as well.