fix(Type): ensure SerializableType updates when assembly name changes
The internal assemblyQualifiedTypeName field within SerializableField
is mutated in the TypePickerAttributeDrawer but when the field data was
changed, it never actually returned a new serialized type when the
ActualType property was called because the setting of the internal
private field wasn't ever checked for updates.

This fix now implements the methods from the
ISerializationCallbackReceiver on the SerializableType to handle
setting the internal data correctly when the class is serialized
and deserialized.

A test has been added for the SerializableType as well.