fix(Attribute): display sensible type picker label for collections
The TypePickerAttributeDrawer would override the label data of
the collection component as soon as it was set to the fully qualified
name of the component, which made the inspector look unsightly.

This fix simply checks to see if the label string matches the selected
type and if it does then just changes the label string to something
more generic and easier to understand.