fix(Type): simplify SerializableType serialization
Instead of handling potentially changed data in the getter of the
actual type this fix ensures changes to `SerializableType` are
serialized by using the callback methods Unity offers for this
purpose. This fix simplifies the change handling but also improves
performance as the actual type can now just be returned directly
instead of having a potentially costly lookup.

The tests for this type have been updated to only test the remaining
actual logic.