fix(Attribute): display sensible type picker label for collections
The TypePickerAttributeDrawer would override the label data of the
collection component as soon as it was set to the fully qualified
name of the component, which made the inspector look unsightly.

This fix simply checks to see if the property being drawn is part of
a collection  and if it does then just changes the label string to
the default label normally used for collection elements.