feat(Velocity): update fields to properties and lists to observable
All usages of public fields have now been switched to properties that
are serialized with Malimbe to create an appropriate backing field to
display in the Unity Inspector.

Any use of a List collection has been updated to use an appropriate
concrete version of the ObservableList, which makes it possible to
track changes occurring in the list at runtime.