feat(Query): update fields to properties and rename pointing to facing
All usage of public fields have now been converted to properties that
are serialized with Malimbe to create an appropriate backing field to
display in the Unity Inspector.

The PointingQuery has also been renamed to FacingQuery as the actual
use of the query is to determine if a GameObject is facing (down it's
local forward direction) another GameObject. The events have also been
updated to reflect this change.

Tests have also been added for both Query types.