fix(Utility): use custom property/backing field on InterfaceContainer
As Malimbe automatically weaves in an `isActiveAndEnabled` check into
any `[Serialized]` property then it would cause an error with the
InterfaceContainer if a type implementing the InterfaceContainer was
displayed in the inspector but it's value was `null`.

This fix removes the automatic weaving of the backing field and
manually creates one to ensure no additional checks are being done.