fix(Utility): set field instead of property after deserialization
The InterfaceContainer was updated to provide a different logic path
when it was deserialized to ensure the `(Field != null)` check wasn't
performed in the `OnAfterDeserialize` call. However, the `Field`
property being set after deserialization is also enough to cause
an error and instead the backing field is what needs setting to
remove the error.