refactor(Extension): use ReadOnlyCollection instead of Collection
The current extension for ICollection only had a single method which
was a read-only operation but any IReadOnlyCollection did not have
access to the ICollection extensions.

The extension is now for IReadOnlyCollection and as it is just a
read-only operation method so far, this means that any ICollection
will have access to the IReadOnlyCollection extension.