feat(Event): update fields to properties and lists to observable
All usages of fields have now been switched to properties that are
serialized with Malimbe to create an appropriate backing field to
display in the Unity Inspector.

All uses of List collections have also been replaced with concrete
versions of the ObservableList, which makes it possible to track
changes occurring to the lists at runtime and provides a single
component that allows mutating the list via UnityEvents.

The BehaviourEnabledObserver has also been re-written to utilise a
Coroutine instead of using InvokeRepeating.

The ObservableList has also had a new ReadOnlyCollection added that
allows for receiving the List collection at any time before Start
is called, however this list is not suitable for using with the
subscribed events or creating any reactive process based on the
List contents.