feat(Modification): remove enable check in GameObjectStateSwitcher
The GameObjectStateSwitcher could automatically switch to a StartIndex
when the component was enabled. However, this functionality shouldn't
be part of the component, instead it is now possible to set the
CurrentIndex of the switched state and then use the
BehaviourEnabledObserver to listen for when the GameObjectStateSwitcher
is enabled then in the Enabled event simply set the CurrentIndex to
the appropriate starting index and call SwitchToCurrentIndex. Even
easier would be to just call SwitchTo(starting-index). The
SwitchToCurrentIndex has been left in to make the API similar to other
APIs where there is a current tracking index.