fix(tests): use UnityTest for ObservableList tests to call start
The ObservableList now requires `Start()` to be called before the list
elements are accessible. This only happens when Play Mode has actually
started and any tests that run purely in editor mode (`[Test]`) won't
have this Start executed.

All tests that now use an ObservableList use UnityTest and all yield
null after the ObservableList has been added to ensure Start has
occurred.

The `wasStartCalled` property in the ObservableList has also been
converted to a field as it makes little sense to have it as a property.