feat(Collection): allow adding unique items to observable list
The ObservableList now has a collection of complimentary methods that
only allow unique elements to be added to the list. This now means the
ObservableSet is obsolete as the ObservableSet was used for unique list
checking. Now, if a unique list is required then simply using the
ObservableList and the Unique methods are enough.