fix(Data): ensure correct list state when emitting inspector events
When changing the size of an ObservableList via the inspector the
custom inspector would not call the events while the list is in the
expected state. The fix is to revert changes Unity already did
through serialization, then call the same API a code change would
call. This ensures the events are raised while the list is in the
same state it would be if the change is done via code or a
UnityEvent listener.