fix(Data): ensure correct list state when setting an element
Setting an element to an existing index basically overrides the
existing element at that index. Thus the ObservableList correctly
raised the Removed and Added event for the previous and new value
respectively. The issue comes from the fact that the SetAt method
only emits the Removed method for the previous element, it doesn't
actually remove it, thus a listener will get told the element was
removed while it is still in the list at that point. This breaks
listeners that are enforcing some kind of unique check over the list
in the event listener methods. The fix is to replace SetAt so it
removes the previous element altogether, then adds the new element
at the same index.