fix(Extension): rename collection index clamp extension method
The GetWrappedAndClamped extension method for IReadOnlyCollections had
a misleading name as it didn't actually wrap the index value if it was
out of bounds. It simply clamps the index value within the bounds of
the collection.

It also allows for a negative index to be provided which starts the
index look up from the back of the collection, but this still isn't
wrapping so the word wrapping has been completely removed to avoid
any confusion.

This ability to allow for a negative index is now via an optional
parameter that is defaulted to do the negative walk.