chore(general): use ObservableList.NonSubscribableElements instead
All components have been updated to use the non-subscribable variant
of the list from ObservableList in case they don't actually handle
changes in a reactive way and instead poll the state of the list.
The tests have also been changed to use the non-subscribable variant
except for the tests that test the reactive components.