fix(Tracking): use correct delta time independent of loop type
`Time.deltaTime` is the preferred way to get the delta time because
it automatically returns the right time depending on whether the
code is running as part of the fixed update loop or the regular
update loop. More information can be found in [Unity's
documentation](https://docs.unity3d.com/ScriptReference/Time-fixedDeltaTime.html).