feat(Process): use interval in seconds instead of frame counter
To make `IProcessable`s framerate independent a time-based solution
is needed. This change replaces the current frame counter based
logic in `MomentProcess` to instead allow setting how often a
process should run based on an interval (in seconds). The next time
the process should run has been implemented as a public property to
allow delaying a process manually when needed.