feat(Process): allow processing independent of the scheduled time
Running a process sometimes is done as a response to something,
rather than continuously in a loop. In these cases having to set the
`NextProcessTime` manually before calling `Process` is an
unnecessary burden to the user. This change allows running a process
at any point in time, while ignoring the scheduled time it should
run next. This method updates the next time to run the process at to
ensure the `Interval` setting is honored.