feat(Tracking): allow GameObject as Target in ObscuranceQuery
ObscuranceQuery currently requires a Collider as the target, however
this is limiting when using objects that contain compound colliders.
This change allows providing a GameObject as the target instead and
use that for the collision check. Additionally an exception is
thrown in case the target doesn't have any Rigidbody, nor Collider
on it to prevent users of the component from silently never getting
any result. More information can be found in [Unity's
documentation](https://docs.unity3d.com/Manual/CollidersOverview.html).