chore(Association): explain why the test results in a warning log
One of the tests for `GameObjectsAssociationActivator` tests whether
an expected warning log message is actually logged. As this warning
log can be seen as a (partial) failure in the test by users running
the tests, the test has been adjusted to log additional information
before and after the expected warning log. These messages explain
that the warning is indeed expected and intended.