fix(Modification): switch from behaviour to component observable list
The ComponentEnabledStateModifier originally used a
SerializedTypeBehaviourObservableList but this only allows picking
classes that extend Behaviour and Renderers in Unity do not extend
Behaviour. Instead, a SerializedTypeComponentObservableList is now used
which allows picking any Behaviours and any Renderers.

The component already ensures it does nothing if a component with no
enabled type is picked.