feat(Collection): add method to set element or add if empty to list
The ObservableList now has a new method that allows setting an element
in the collection or adding it to the collection if the collection is
empty.

This is a somewhat obscure requirement and it in fact does two
operations but is required due to the limitations of UnityEvents where
only rudimentary logic is possible.

The Remove method has also been updated so it returns a bool and a
DoRemove method has been added that returns nothing so it can still
be used in UnityEvents.