feat(Collection): provide default value for Clear parameter
The ObservableList Clear method now has a default value for the Clear
parameter set to false to just reduce the amount of times Clear(false)
is required, when simply Clear() can be used, which is the standard
operation.

The method parameter has also been renamed so false is the obvious
default parameter and the internal usage doesn't have to be a negation
of the value to perform the operation.