feat(Extension): ability to queue a behaviour action to run on enable
The BehaviourExtensions provide a mechanism to queue up a behaviour
action on a component to run when the component becomes enabled. This
makes it possible to run methods on disabled components by waiting
for them to become enabled and then running the appropriate method.

This is very useful for calling ObservableList methods such as Add
in the OnEnable method of other components in case there is a race
condition of which component becomes enabled first.