fix(Collision): ensure ignored collisions are resumed on disable
The CollisionIgnorer was not resuming collisions when the component
was disabled due to a short circuit check failing because the logic
was there to prevent resuming collisions when removing an item from
the list.

The fix ensures that check is only done when the toggle is done
via the list mutation and not when disabled.

An additional parameter has also been added to optionally toggle
collisions on inactive colliders if required.