feat(Cast): provide validity data along with hit data
The PointsCast now allows target collisions that do not pass the
validity rules to emit their data instead of null but along with a
new IsValid flag so it's possible to know whether the collision is
with a valid target.

This data is now used in the ObjectPointer so it's possible with the
ObjectPointer to know when you're hovering over an invalid target
but selecting an InvalidTarget still emits `Selected(null)` as an
invalid target is considered the same as no target at all.