fix(Action): prevent null actions being subscribed to/from
There was an issue when increasing or decreasing the Sources List size
in the inspector as this would trigger the OnChange handlers for the
Sources and attempt to subscribe/unsubscribe on the Action. However, by
default Unity will create a null action when increasing the size of an
empty List which would then attempt to register a null Action and
therefore it would throw an error.

The larger problem became when then trying to remove the null actions
in the inspector as because errors were being thrown meant the
inspector wouldn't update and would stay in a broken state.

This fix ensures the passed action is null checked. This does mean that
null actions will silently fail and not provide any useful feedback if
attempted, but it's required to work around the limitations and issues
in the Unity editor.