fix(Visual): ensure points renderer utilizes source repeated segment
Previously, The PointsRenderer component would use the RepeatedSegment
GameObject as a template to instantiate the repeated clones for the
line to draw, however this meant that the actual RepeatedSegment
GameObject wasn't being used in the render process.

This differs from both the Start and End object, which are both used
plus it also meant that the original RepeatedSegment GameObject would
be visible at a position of `0,0,0` and could be visible in scene.

This fix ensures the segment clones are always one less than required
so the actual RepeatedSegment can be used as the first object to
display.