fix(Pointer): display invalid objects when target is not valid
The change introduced in 050cfc6 that meant all cast collisions now
contain HitData even if the target is invalid meant that the current
check in the Pointer was not enough to know if the target was invalid.

The new `IsValid` flag on the cast data is now also checked to see
if the target is valid or invalid.