fix(general): reduce garbage-collected allocations in editor
Component lookups create garbage-collected allocations in case there
exists no component and the lookup is running during design time (in
the Editor). This change prevents the allocations by looking up all
components instead and returning the first one if any is found.
While this behavior is the same as a regular component lookup, the
lookup method call itself has to change to an extension one.