feat(Utility): simplify interface container drawers
The custom PropertyDrawer for `InterfaceContainer` or any subclass
of that type has been updated to be simpler. The only functionality
it offers now is to help with assigning instances. While the editor
allows to easily pick a game object, picking a specific component is
more difficult. The PropertyDrawer has been changed to simplify the
latter scenario as one can now drag and drop a specific GameObject
or Component onto the ObjectField which then optionally will show a
popup window that allows selecting one of the components on the
selected game object. The same logic applies to picking a game
object with the ObjectField itself, without dragging. A component is
said to be "matching" if the type matches the field's type, which is
used early, to even prevent dropping an invalid component onto the
ObjectField.