fix(Pointer): default values of AllocationFreeReadOnlyList
The recently merged addition of the new type
`AllocationFreeReadOnlyList` comes with unresolved issues. This
change fixes the usage in `ObjectPointer`: Since the new type is a
value type it's not nullable. Additionally the type itself has been
changed to handle creations of the type via the implicit default
constructor the C# language offers for any value type. Since that
constructor just initializes the value with "zero" for each field
the underlying `List` will be null. A fallback to treat this case
the same as an empty list successfully handles this case.