perf(Cast): use non-allocating physics queries
The `Physics` API allows querying the physics system without doing
any new (heap) allocations. As physics queries are often run each
frame (or even more often, depending on the fixed time step) this
change improves performance by ensuring a single array of results is
used. By making this array fixed in size it is now allocated just
once on startup and is then used in all queries. Because the array
is potentially bigger than the actual results of a query the API has
been changed to return an `ArraySegment` instead. Callers of this
API need to use the copy API available on the `ArraySegment` type if
they intend to mutate the returned array (found inside the segment)
as it will be reused for other queries. For the rare case of a
caller being unable to only handle the returned segment, instead of
the whole underlying array, a method has been added to allow
clearing the cached array. The `SurfaceLocator` has been changed to
use this new method as it is unable to sort a segment due to
limitations of the `Array.Sort` API (no overload exists that allows
passing a concrete `Comparison` instead of a boxed `IComparer` that
*also* offers a start and count parameter).