perf(Tracking): replace boxed IComparer with alloc-free Comparison
When sorting the interface `IComparer` is used, but this is giving
heap allocations due to boxing in the `Sort` API calls. To eliminate
these heap allocations the implementations have been updated to
instead pass the comparing method directly. Passing method pointers
in C# means the pointer gets boxed, too, but this is now only done
once, reducing the allocations to a minimum.