feat(Type): get Vector3 representation from Vector3State
A new ToVector3 method has been added to the Vector3State component
which returns the state of the Vector3State as a Vector3 representation
which allows for obtaining useful multiplication data based on the
current state. For example, if the state is used to determine valid
axes then a state of `false, true, false` could represent an interest
in a `y axis` so obtaining the Vector3 equivalent of this would return
`0f, 1f, 0f` which could be used to multiply against another value to
only consider the y axis valid.