feat(Transformation): aggregate collection on existing data set
Currently the Collection Aggregators require the Transform method to
be called with a value to apply to the current index and then be
transformed. However, it may be required to simply transform the
pre-existing collection that has already been set up.

This new `Transform()` method allows for that by simply passing the
value at the current index to the Aggregator which will in turn simply
set that value back at the current index and then perform the transform
on the collection.