fix(Extension): prevent exception if behaviour is destroyed
There was an issue with the BehaviourExtensions RunWhenActiveAndEnabled
method that it would throw an exception if the behaviour was destroyed
whilst the OnBeforeRender was running, for example if a scene was
switched.

This fix simply checks to see if the behaviour is null in the
OnBeforeRender callback and if it is then removes the listener from
the event and exits the method.