feat(Collision): emit when active collision is added or removed
The ActiveCollisionsContainer now has events that emit when a new
active collision occurs and when an existing collision ceases. This
extends the functionality that previously happened where an event
was emitted on the first collision and on the last final collision
ceasing.

The event data payloads have also been changed for the events as
there is no point in emitting the entire collision collection for the
first collision as it will only contain one collision, so it may as
well just emit that single collision. The AllStopped event also now
does not emit anything as when all collisions have ceased then no
collision data will be available to emit anyway so it was just emitting
an empty collection.

A fix to the validity check has also been added to ensure it retrieves
the collider containing transform in case of compound colliders.