fix(Modification): create new TransformData when setting by GameObject
The TransformPropertyApplier `SetSource(GameObject source)` method was
not actually performing the correct logic, when setting the `Source`
via a GameObject, it actually should create a brand new TransformData
from that source GameObject and then set the `Source` property, which
internally performs the relevant caching.

Without this, the cached object would be set up correctly but the
public Source property would be incorrect. This method is a creation
method and is rarely called and therefore any overhead is acceptable.