feat(Modification): allow dynamic transition destination
The TransformPropertyApplier destination location for the
transformation can now be specified as a dynamic destination meaning
that any changes to the destination (the `Source` property) that
occur during the transition routine will be applied to the transition
process.

This is useful if the `Source` properties change during the transition
and the `Target` should move to the updated properties instead of the
original specified destination.