fix(Event): do not mutate data when restrictable proxy is invalid
The RestrictableSingleEventProxyEmitter would still mutate the Payload
data in the ProxyEmitter even if the validity checks failed. This is
not correct behaviour as if something does not pass validity checks
then it should cease to operate and not change any data.

The fix is to ensure the SingleEventProxyEmitter checks the valid
state and if it is not valid then it should reset the payload back
to what it was before the payload was received.

The Payload still needs setting before the validity check is done
because other classes rely on a Payload being available to achieve
the validity check.