fix(Modification): only use source cache when setting by GameObject
The TransformPropertyApplier was using the source cache internal field
for inline operations but this meant that any changes to the public
Source TransformData property would not be reflected in the internal
code logic as it was never updating the cache unless the Source
property was changed.

This fix removes the source cache from being used inline and always
uses the public Source property. The source cache is now only used
for setting Source with a GameObject via the `SetSource` method.

The use of the source cache here removes the need for creating a
new instance of TransformData from the GameObject and therefore removes
garbage creation.