fix(Tracking): reset distance check when pinch points change
The PinchScaler was not reporting the correct distances between
pinch points if either the primary or secondary point changed
(or was cleared) as the code would only reset the previous
distance back to null if the component was disabled. It should
also reset the previous distance check to null if either the
primary or secondary points change so the previous distance in
the process calculation is recreated to have the correct initial
distance delta.