fix(Pointer): ensure element visibility allows pointer to function
The pointer element visibility was causing the pointer functionality
to break if certain elements were set to hidden. This is because the
element container would simply be turned off yet the element GameObject
would still be needed in the scene to work.

This has been fixed now by having an element container and an element
mesh container and the mesh container is what is turned on/off by
the visibility settings whereas the element container is always on
when the pointer is active so pointer functionality still works
correctly.