feat(Pointer): rehandle existing pointer data
There are occasions where the existing pointer data needs reprocessing
but the component requiring the rehandling does not contain pointer
data.

One case is when the PointerElement visibility changes, it requires
the existing pointer data to be reprocessed again.

This method provides a quick and simple way to reprocess the existing
pointer data.