feat(Rule): match if behaviour is found an is enabled
The AnyBehaviourEnabledRule matches if any specified behaviour is found
on the given GameObject and whether the behaviour is enabled. If the
behaviour is not found or is found but disabled then the rule will not
match.

This is useful to disable rules applied to GameObjects if the specified
Behaviour disabled.