fix(Rule): ensure layer rule matches correctly against layer mask
The AnyLayerRule was not correctly matching against the specified
LayerMask and the tests were also incorrect and therefore not
highlighting the issue.

In the tests where it would set up the LayerMask by doing
`subject.LayerMask = LayerMask.NameToLayer("UI")` this would not
actually set the LayerMask property to `UI` instead it would set
it to `Default` and `Ignore Raycast`. The correct way to set the
LayerMask property is to use `LayerMask.GetMask("UI")` which then
will select the correct property and highlight the issue in the
AnyLayerRule.

The AnyLayerRule is then fixed by checking the LayerMask property
is equal to the existing LayerMask with the target layer pushed
on to it.

This makes all of the tests pass and also works by testing with
scene objects.