fix(Collision): utilize both colliding objects in equality check
The CollisionNotifier.EventData equality check was only checking to
see if one side of the collision matched to determine if the collision
data was equal.

This is incorrect as the following should be true:

```
A collides with B == A collides with B
A collides with B != A collides with C
A collides with B != C collides with B
```

However, previously this was actually the case

`A collides with B == C collides with B`

Because the second object of the collision was only being checked for
equality therefore if two different objects (A, C) collided with the
same target (B) then the EventData would be considered equal whereas
the collision is not equal.

The fix is to ensure if there is a forwarding source of the collision
then that should also be used in the equality check.

Additional tests have been added to test the functionality of the
CollisionNotifier and CollisionTracker and to test this EventData.