fix(Tracking): use injected scale for located surface output
The SurfaceLocator currently only outputs positional information
about any located surface and therefore it overrides the rotation
data output with the rotation data of the given input.

This should also be done with the scale in case anything digesting
surface data is also modifying scale data too as it would receive
scale data of the transform the collision occurred with.