feat(Collision): ability to ignore tracked collision states
The CollisionNotifier now has an enum flag to determine which collision
states to process. Any ignored collision state simply will return early
in the OnCollision/Trigger method related to the state.

This gives the option of forcing any CollisionNotifier to never process
a state if it's never required because at the moment it will always
at least run through the process of looking for CollisionNotifiers.

The CollisionTracker also directly utilizes the new property to
immediately exit out of the Unity collision checks just to prevent
any additional unnecessary work such as setting the event data.