fix(Tracking): prevent duplicate active collisions in collection
The ActiveCollisionsContainer could have duplicate collisions added
to the container which is incorrect because only 1 collision is really
valid and required. It would also create a leak if the collision was
never removed but due to a Unity bug if the colliding object was
disabled then the collision would not be removed but would be readded
again if the colliding object was enabled again.