feat(Tracking): terminate surface locator on certain collisions
A new `LocatorTermination` Rule has been added to the SurfaceLocator
that allows a given rule to determine if the locator hits a target
which matches the rule to simply terminate the locator operation and
no longer continue to search for a valid surface.

This is in addition to the `TargetValidity` Rule which will ignore
invalid targets but continue searching for the next valid target. The
`LocatorTermination` Rule will prevent this search from continuing.