fix(Collision): stop collisions when either side is disabled
There is an issue in Unity where when two objects are colliding will
create a collision started event, but whilst the objects are
intersecting, if one of the objects is disabled then Unity does not
consider this to stop the collision between the two objects so no
OnCollisionExit or OnTriggerExit occurs. However, if the disabled
object is re-enabled and is still in the same intersecting position
then Unity will consider this a new Enter event.

This causes a problem for the CollisionTracker as it means the state
of the tracker can become out of sync. If a collision starts twice
but no stop was called between, then the tracked collision state is
not valid. This would then have knock on effects in usages such as
the ActiveCollisionsContainer where a started collision is treated
as an active collision, but then if the object is disabled then
no stopped collision occurs and therefore that would still be
considered as an active collision.

To fix this issue and work around the shortcomings of Unity, a new
component is added at runtime to the GameObject that contains the
collider that is colliding with the CollisionTracker. This new
component then monitors the disabled state of the GameObject and upon
that GameObject becoming disabled will stop the collision manually
on the linked CollisionTracker. It will also then remove any traces
of the magically created component.

The CollisionTracker now also tracks any colliders it may be
intersecting with and therefore if the CollisionTracker is disabled
then any collider it may be linked to also has the Collision Stopped
event emitted for to ensure both sides of the collision adhere to the
disabled state.

This new functionality is optional and can be turned off by the `Stop
Collisions On Disable` property, which can only be set at edit time to
prevent fundamental state getting out of sync if it was changed at
runtime.

The `AddComponent` and `Destroy` are used over a pooling method because
these only occur on the `Enter/Exit` events so are not firing all the
time and therefore performance hits should be reduced. But also,
pooling would create a large set of GameObjects for each
CollisionTracker and they would all have to independently manage these
pools themselves resulting in a potential large number of pooled
GameObjects being created on scene start, which would have a potential
much larger impact on initial performance.

The tests have been updated to showcase the issue of disabling a
GameObject on either side of the collision now can successfully result
in the Stopped event being emitted.