fix(Collision): perform equality check on collider and forward source
The CollisionNotifier.EventData was performing the equality check on
the ForwardSource of the collision and the containing Transform of
the collider being intersected. This actually can report invalid
equality data as two compound colliders being touched by the same
ForwardSource would report as being the same Collision Data which
they're actually not.

The fix is to just do the equality check directly on the Collider and
the ForwardSource.

The HashCode of the EventData should also be a combination of the
two equality checked objects to ensure its uniqueness.