feat(Extension): get collider attached rigidbody even when inactive
The default `Collider.attachedRigidbody` does not report the parent
Rigidbody for a Collider if the GameObject (or parent GameObject)
is inactive, instead it returns null.

This isn't helpful when attempting to deal with stopping collisions
when the GameObject is being disabled. The solution is to provide
a new extension method that uses the `Collider.attachedRigidbody`
getter by default but if the collider GameObject is not active then
it falls back to using a more brute force approach of getting
all Rigidbodies in the parent GameObjects and returning the first
one found.